Algumas correções em Customização e Vendas

+++ VendaDeAutomoveis/Entidades/Roda.cs
@@ -12,6 +12,6 @@ namespace VendaDeAutomoveis.Entidades
 
         public int Aro { get; set; }
         
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
     }
 }
